+++ packages/react-ui/src/components/atoms/LabelWithIndicator.js
@@ -1,34 +1,30 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import makeStyles from '@mui/styles/makeStyles';
+import { styled } from '@mui/material/styles';
+
 import Typography from './Typography';
 
-const useStyles = makeStyles((theme) => ({
-  indicator: {
-    marginLeft: theme.spacing(0.5),
-
-    '.Mui-disabled &': {
-      color: theme.palette.text.disabled
-    }
+const LabelIndicator = styled(Typography)(({ theme }) => ({
+  marginLeft: theme.spacing(0.5),
+  '.Mui-disabled &': {
+    color: theme.palette.text.disabled
   }
 }));
 
 const LabelWithIndicator = ({ label, type }) => {
-  const classes = useStyles();
   const isRequired = type === 'required';
 
   return (
     <>
       {label}
-      <Typography
+      <LabelIndicator
         component='span'
         variant='inherit'
         color='textSecondary'
         weight='regular'
-        className={classes.indicator}
       >
         {isRequired ? '(required)' : '(optional)'}
-      </Typography>
+      </LabelIndicator>
     </>
   );
 };
@@ -37,7 +33,7 @@ LabelWithIndicator.defaultProps = {
   type: 'optional'
 };
 LabelWithIndicator.propTypes = {
-  label: PropTypes.string.isRequired,
+  label: PropTypes.oneOfType([PropTypes.element, PropTypes.string]).isRequired,
   type: PropTypes.oneOf(['optional' | 'required'])
 };