+++ packages/react-ui/storybook/stories/atoms/Paper.stories.js
@@ -21,7 +21,7 @@ const options = {
       url: 'https://www.figma.com/file/nmaoLeo69xBJCHm9nc6lEV/CARTO-Components-1.0?node-id=1447%3A38722'
     },
     status: {
-      type: 'needUpdate'
+      type: 'needsUpdate'
     }
   }
 };