+++ packages/react-ui/storybook/stories/atoms/Slider.stories.js
@@ -135,7 +135,7 @@ const options = {
       url: 'https://www.figma.com/file/nmaoLeo69xBJCHm9nc6lEV/CARTO-Components-1.0?node-id=1534%3A32732'
     },
     status: {
-      type: 'needUpdate'
+      type: 'needsUpdate'
     }
   }
 };