+++ packages/react-ui/storybook/stories/organisms/AppBar.stories.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Avatar, Chip, Grid, IconButton, Menu, MenuItem } from '@mui/material';
 import { CheckCircleOutline } from '@mui/icons-material';
-import AppBar from '../../../src/components/organisms/AppBar';
+import AppBar from '../../../src/components/organisms/AppBar/AppBar';
 import Typography from '../../../src/components/atoms/Typography';
 
 const options = {
@@ -34,7 +34,7 @@ const options = {
       url: 'https://www.figma.com/file/nmaoLeo69xBJCHm9nc6lEV/CARTO-Components-1.0?node-id=3121%3A61435&t=HI1Xd9yGuft4g9kq-0'
     },
     status: {
-      type: 'readyToReview'
+      type: 'validated'
     }
   }
 };